Test.ThreadNet.General

Documentation

data PropGeneralArgs blk

Constructors

PropGeneralArgs

Fields

pgaBlockProperty ∷ blk → Property

test if the block is as expected

For example, it may fail if the block includes transactions that should have expired before/when the block was forged.

pgaCountTxs ∷ blk → Word64

the number of transactions in the block

pgaExpectedCannotForge ∷ SlotNo → NodeId → WrapCannotForge blk → Bool

whether this CannotForge was expected

pgaFirstBlockNo ∷ BlockNo

the block number of the first proper block on the chain

At time of writing this comment... For example, this is 1 for Byron tests and 0 for mock tests. The epoch boundary block (EBB) in slot 0 specifies itself as having block number 0, which implies the genesis block is block number 0, and so the first proper block is number 1. For the mock tests, the first proper block is block number 0.

pgaFixedMaxForkLength ∷ Maybe NumBlocks

the maximum length of a unique suffix among the final chains

If not provided, it will be crudely estimated. For example, this estimation is known to be incorrect for PBFT; it does not anticipate PBftExceededSignThreshold.

pgaFixedSchedule ∷ Maybe LeaderSchedule

the leader schedule of the nodes

If not provided, it will be recovered from the nodes' Tracer data.

pgaSecurityParam ∷ SecurityParam

pgaTestConfig ∷ TestConfig

pgaTestConfigB ∷ TestConfigB blk

calcFinalIntersectionDepth ∷ ∀ blk. HasHeader blk ⇒ PropGeneralArgs blk → TestOutput blk → NumBlocks

What was the most number of signed blocks needed to be dropped from a final chain in order to reach the final chains' common prefix?

NOTE: This count excludes EBBs.

prop_general ∷ ∀ blk. (Condense blk, Condense (HeaderHash blk), Eq blk, RunNode blk) ⇒ PropGeneralArgs blk → TestOutput blk → Property

prop_general_semisync ∷ ∀ blk. (Condense blk, Condense (HeaderHash blk), Eq blk, RunNode blk) ⇒ PropGeneralArgs blk → TestOutput blk → Property

Like prop_general but instead assuming semi-synchrony

For now, this simply disables a few Propertys that depend on synchrony.

prop_inSync ∷ ∀ blk. HasHeader blk ⇒ TestOutput blk → Property

All final chains have the same block number

runTestNetwork ∷ ∀ blk. (RunNode blk, TxGen blk, TracingConstraints blk, HasCallStack) ⇒ TestConfig → TestConfigB blk → (∀ m. IOLike m ⇒ TestConfigMB m blk) → TestOutput blk

Thin wrapper around runThreadNetwork

TestConfig

data TestConfig

Test configuration that does not depend on the block

The primary motivation for separating this type from TestConfigB and TestConfigMB is so that the instance Arbitrary TestConfig can be reused by multiple tests using different blks: as of yet, no block (each of which realizes a ledger-protocol combination) influences the validity of these data.

Constructors

TestConfig

Fields

initSeed ∷ Seed

nodeTopology ∷ NodeTopology

numCoreNodes ∷ NumCoreNodes

numSlots ∷ NumSlots

TODO generate in function of k

data TestConfigB blk

Test configuration that depends on the block (incl the ledger and/or protocol) but not on the monad

Some fields do not explicitly involve the blk type, but their semantics (at least their validity) does depend on the semantics of the underlying ledger and/or protocol. For example, nodeJoinPlan is here instead of in TestConfig because different blocks can withstand different degrees of absence/lateness. And epochSize is here because eg the Byron ledger assumes a fixed epoch size of 10k. And so on.

Constructors

TestConfigB

Fields

forgeEbbEnv ∷ Maybe (ForgeEbbEnv blk)

future ∷ Future

messageDelay ∷ CalcMessageDelay blk

nodeJoinPlan ∷ NodeJoinPlan

nodeRestarts ∷ NodeRestarts

txGenExtra ∷ TxGenExtra blk

version ∷ (NodeToNodeVersion, BlockNodeToNodeVersion blk)

data TestConfigMB m blk

Test configuration that depends on the block and the monad

The primary motivation for separating this type from TestConfigB is so that TestConfigB can occur in contexts (such as in PropGeneralArgs) for which the m parameter is irrelevant and hence unknown.

Constructors

TestConfigMB

Fields

nodeInfo ∷ CoreNodeId → TestNodeInitialization m blk

mkRekeyM ∷ Maybe (m (RekeyM m blk))

runTestNetwork immediately runs this action once in order to initialize an RekeyM value that it then reuses throughout the test

truncateNodeJoinPlan ∷ NodeJoinPlan → NumCoreNodes → (NumSlots, NumSlots) → NodeJoinPlan

truncateNodeRestarts ∷ NodeRestarts → NumSlots → NodeRestarts

truncateNodeTopology ∷ NodeTopology → NumCoreNodes → NodeTopology

Expected CannotForge

noExpectedCannotForges ∷ SlotNo → NodeId → WrapCannotForge blk → Bool

Expect no CannotForges

Re-exports

data ForgeEbbEnv blk

How to forge an EBB

Constructors

ForgeEbbEnv

Fields

forgeEBB ∷ TopLevelConfig blk → SlotNo → BlockNo → ChainHash blk → blk

data TestOutput blk

Constructors

TestOutput

Fields

testOutputNodes ∷ Map NodeId (NodeOutput blk)

testOutputTipBlockNos ∷ Map SlotNo (Map NodeId (WithOrigin BlockNo))

noCalcMessageDelay ∷ CalcMessageDelay blk

plainTestNodeInitialization ∷ ProtocolInfo m blk → TestNodeInitialization m blk