Test.ThreadNet.Network

Description

Setup network

Documentation

newtype CalcMessageDelay blk

Compute the chain diffusion delay

This is the number of slots a MsgRollForward should arrive after the forge slot of the header it is carrying.

It may depend on the (sender, recipient), the current slot, and header.

Constructors

CalcMessageDelay ((CoreNodeId, CoreNodeId) → SlotNo → Header blk → NumSlots)

data ForgeEbbEnv blk

How to forge an EBB

Constructors

ForgeEbbEnv

Fields

forgeEBB ∷ TopLevelConfig blk → SlotNo → BlockNo → ChainHash blk → blk

type RekeyM m blk

Arguments

= CoreNodeId

→ ProtocolInfo m blk

→ SlotNo
The slot in which the node is rekeying

→ (SlotNo → m EpochNo)
Which epoch the slot is in

→ m (TestNodeInitialization m blk)
tniProtocolInfo should include new delegation cert/operational key, and tniCrucialTxs should include the new delegation certificate transaction

How to rekey a node with a fresh operational key

When there is a NodeRekey scheduled in the NodeRestarts, the test node will restart and use tnaRekeyM to compute its new ProtocolInfo.

data TestNodeInitialization m blk

Data used when starting/restarting a node

Constructors

TestNodeInitialization

Fields

tniCrucialTxs ∷ [GenTx blk]

these transactions are added immediately and repeatedly (whenever the ledgerTipSlot changes)

In particular, a leading node's crucial transactions must (if valid) enter its mempool each slot before the node takes the mempool snapshot that determines which transactions will be included in the block it's about to forge.

tniProtocolInfo ∷ ProtocolInfo m blk

data ThreadNetworkArgs m blk

Parameters for the test node net

Constructors

ThreadNetworkArgs

Fields

tnaForgeEbbEnv ∷ Maybe (ForgeEbbEnv blk)

tnaFuture ∷ Future

tnaJoinPlan ∷ NodeJoinPlan

tnaNodeInfo ∷ CoreNodeId → TestNodeInitialization m blk

tnaNumCoreNodes ∷ NumCoreNodes

tnaNumSlots ∷ NumSlots

tnaMessageDelay ∷ CalcMessageDelay blk

tnaSeed ∷ Seed

tnaMkRekeyM ∷ Maybe (m (RekeyM m blk))

tnaRestarts ∷ NodeRestarts

tnaTopology ∷ NodeTopology

tnaTxGenExtra ∷ TxGenExtra blk

tnaVersion ∷ NodeToNodeVersion

tnaBlockVersion ∷ BlockNodeToNodeVersion blk

type TracingConstraints blk = (Show blk, Show (ApplyTxErr blk), Show (Header blk), Show (GenTx blk), Show (Validated (GenTx blk)), Show (GenTxId blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), Show (CannotForge blk), HasNestedContent Header blk)

noCalcMessageDelay ∷ CalcMessageDelay blk

plainTestNodeInitialization ∷ ProtocolInfo m blk → TestNodeInitialization m blk

runThreadNetwork ∷ ∀ m blk. (IOLike m, MonadTime m, MonadTimer m, RunNode blk, TxGen blk, TracingConstraints blk, HasCallStack) ⇒ SystemTime m → ThreadNetworkArgs m blk → m (TestOutput blk)

Setup a network of core nodes, where each joins according to the node join plan and is interconnected according to the node topology

We run for the specified number of blocks, then return the final state of each node.

Tracers

data MiniProtocolFatalException

Any synchronous exception from a directedEdge

Constructors

MiniProtocolFatalException

Fields

mpfeType ∷ !TypeRep

Including the type explicitly makes it easier for a human to debug

mpfeExn ∷ !SomeException

mpfeClient ∷ !CoreNodeId

mpfeServer ∷ !CoreNodeId

data MiniProtocolState

Constructors

MiniProtocolDelayed

MiniProtocolRestarting

Test Output

data NodeDBs db

A vector with an element for each database of a node

The db type parameter is instantiated by this module at types for mock filesystems; either the MockFS type or reference cells thereof.

Constructors

NodeDBs

Fields

nodeDBsImm ∷ db

nodeDBsVol ∷ db

nodeDBsLgr ∷ db

data NodeOutput blk

Constructors

NodeOutput

Fields

nodeOutputAdds ∷ Map SlotNo (Set (RealPoint blk, BlockNo))

nodeOutputCannotForges ∷ Map SlotNo [CannotForge blk]

nodeOutputFinalChain ∷ Chain blk

nodeOutputFinalLedger ∷ LedgerState blk

nodeOutputForges ∷ Map SlotNo blk

nodeOutputHeaderAdds ∷ Map SlotNo [(RealPoint blk, BlockNo)]

nodeOutputInvalids ∷ Map (RealPoint blk) [ExtValidationError blk]

nodeOutputNodeDBs ∷ NodeDBs MockFS

nodeOutputSelects ∷ Map SlotNo [(RealPoint blk, BlockNo)]

nodeOutputUpdates ∷ [LedgerUpdate blk]

data TestOutput blk

Constructors

TestOutput

Fields

testOutputNodes ∷ Map NodeId (NodeOutput blk)

testOutputTipBlockNos ∷ Map SlotNo (Map NodeId (WithOrigin BlockNo))