Test.ThreadNet.Util.NodeJoinPlan

Node Join Plan

newtype NodeJoinPlan

In which slot each node joins the network

Constructors

NodeJoinPlan (Map CoreNodeId SlotNo)

coreNodeIdJoinSlot ∷ HasCallStack ⇒ NodeJoinPlan → CoreNodeId → SlotNo

Partial; error for a node not in the plan

genNodeJoinPlan

Arguments

∷ NumCoreNodes
PRECONDITION: non-negative

→ NumSlots
PRECONDITION: positive

→ Gen NodeJoinPlan

Generate a NodeJoinPlan consistent with the given properties

INVARIANT: Nodes with higher Ids will not join before nodes with lower Ids. This eliminates some uninteresting symmetry and makes the counter-examples easier for humans to interpret.

nodeIdJoinSlot ∷ HasCallStack ⇒ NodeJoinPlan → NodeId → SlotNo

Partial; error for a node not in the plan

shrinkNodeJoinPlan ∷ NodeJoinPlan → [NodeJoinPlan]

Shrink a node join plan

INVARIANT no inter-join delay increases

Specifically, we shrink by setting some of the delays to 0.

trivialNodeJoinPlan ∷ NumCoreNodes → NodeJoinPlan

All nodes join immediately