Test.Util.ChainDB

Documentation

data MinimalChainDbArgs m blk

Minimal set of arguments for creating a ChainDB instance for testing purposes.

Constructors

MinimalChainDbArgs

Fields

mcdbTopLevelConfig ∷ TopLevelConfig blk

mcdbChunkInfo ∷ ChunkInfo

Specifies the layout of the ImmutableDB on disk.

mcdbInitLedger ∷ ExtLedgerState blk

The initial ledger state.

mcdbRegistry ∷ ResourceRegistry m

Keeps track of non-lexically scoped resources.

mcdbNodeDBs ∷ NodeDBs (StrictTVar m MockFS)

File systems underlying the immutable, volatile and ledger databases. Would be useful to default this to StrictTVar's containing empty MockFS's.

data NodeDBs db

A vector with an element for each database of a node

The db type parameter is instantiated by this module at types for mock filesystems; either the MockFS type or reference cells thereof.

Constructors

NodeDBs

Fields

nodeDBsImm ∷ db

nodeDBsVol ∷ db

nodeDBsLgr ∷ db

emptyNodeDBs ∷ MonadSTM m ⇒ m (NodeDBs (StrictTVar m MockFS))

fromMinimalChainDbArgs ∷ (MonadThrow m, MonadSTM m, ConsensusProtocol (BlockProtocol blk)) ⇒ MinimalChainDbArgs m blk → ChainDbArgs Identity m blk

Creates a default set of of arguments for ChainDB tests.

mkTestChunkInfo ∷ LedgerConfig blk ~ EraParams ⇒ TopLevelConfig blk → ChunkInfo

Utility function to get a default chunk info in case we have EraParams available.