Test.Util.ChainUpdates

Description

Generate sequences of updates to model an evolving chain

Documentation

data ChainUpdate

Constructors

AddBlock TestBlock

SwitchFork (Point TestBlock) [TestBlock]
Roll back to the given Point, and then immediately roll forward by the given TestBlocks.

data UpdateBehavior

Different strategies how to generate a sequence of ChainUpdates.

Constructors

SelectedChainBehavior
Chain updates tracking the selected chain of an honest node. In particular, this includes:

All blocks involved are valid.

Every ChainUpdate improves the chain.

TentativeChainBehavior
Chain updates tracking the tentative chain of an honest node (in the context of diffusion pipelining). This is similiar to SelectedChainBehavior, but allows for the following sequence of ChainUpdates:

AddBlock blk for blk invalid

SwitchFork (prevPoint blk) [blk'] where blk' is preferable to blk.

genChainUpdates

Arguments

∷ UpdateBehavior

→ SecurityParam

→ Int
The number of updates to generate

→ Gen [ChainUpdate]

toChainUpdates ∷ [ChainUpdate] → [ChainUpdate TestBlock TestBlock]

Tests

prop_genChainUpdates ∷ SecurityParam → Int → Property

Test that applying the generated updates gives us the same chain as cusCurrentChain.