Test.Util.Classify

Description

Classification of lists of symbols

Intended for qualified import.

import qualified Test.Util.Classify as C

Documentation

data Predicate a b

Predicate over a list of as, using classification b

Constructors

Predicate

Fields

predApply ∷ a → Either b (Predicate a b)

Given an a, either successfully classify as b or continue looking

predFinish ∷ Maybe b

End of the string

The predicate is given a final chance to return a value.

classify ∷ ∀ a b. [Predicate a b] → [a] → [b]

Do a linear scan over the list, returning all successful classifications

maximum ∷ ∀ a b. Ord b ⇒ (a → Maybe b) → Predicate a b

Maximum value found, if any

predicate ∷ (a → Either b (Predicate a b)) → Predicate a b

Construct simply predicate that returns Nothing on termination

Example

data Tag

example ∷ [Int] → [Tag]