Test.Util.FS.Sim.FsTree

Description

Internal part of the mock file system

Intended for qualified import

import Test.Util.FS.Sim.FsTree (FsTree)
import Test.Util.FS.Sim.FsTree as FS

FsTree type and indexing functions

data FsTree a

Simple in-memory representation of a file system

Constructors

File !a

Folder !(Folder a)

data FsTreeError

File access error

Constructors

FsExpectedDir FsPath (NonEmpty Text)
A path ..a.. where a is a file rather than a dir

We record both the full path and the invalid suffix.

FsExpectedFile FsPath
A path ..a.. where a is a dir rather than a file

No suffix is specified (it must be the last part of the file)

FsMissing FsPath (NonEmpty Text)
A path ..a.. or ../a where directory or file a is missing

We record both the full path and the missing suffix.

FsExists FsPath
A file was opened with the O_EXCL flag, but it already existed.

example ∷ Monoid a ⇒ FsTree a

Example

Construction

empty ∷ FsTree a

Indexing

getDir ∷ FsPath → FsTree a → Either FsTreeError (Folder a)

getFile ∷ FsPath → FsTree a → Either FsTreeError a

index ∷ FsPath → FsTree a → Either FsTreeError (FsTree a)

Index the FsTree by the given FsPath.

File system operations

createDirIfMissing ∷ FsPath → FsTree a → Either FsTreeError (FsTree a)

Create a directory if it does not already exist

createDirWithParents ∷ FsPath → FsTree a → Either FsTreeError (FsTree a)

Create a directory and its parents if they do not already exist

openFile ∷ Monoid a ⇒ FsPath → AllowExisting → FsTree a → Either FsTreeError (FsTree a)

Open a file: create it if necessary or throw an error if it existed already wile we were supposed to create it from scratch (when passed MustBeNew).

removeFile ∷ FsPath → FsTree a → Either FsTreeError (FsTree a)

Remove a file (which must exist)

renameFile ∷ FsPath → FsPath → FsTree a → Either FsTreeError (FsTree a)

Rename the file (which must exist) from the first path to the second path. If there is already a file at the latter path, it is replaced by the new one.

replace ∷ FsPath → a → FsTree a → Either FsTreeError (FsTree a)

Replace the contents of the specified file (which must exist)

Pretty-printing

pretty ∷ ∀ a. (a → String) → FsTree a → String