Test.Util.LogicalClock

Description

Logical time (in terms of abstract " ticks ")

Intended for qualified import

import Test.Util.LogicalClock (LogicalClock)
import qualified Test.Util.LogicalClock as LogicalClock

API

data LogicalClock m

Logical clock (in terms of ticks rather than actual UTCTime)

Constructors

LogicalClock

Fields

getCurrentTick ∷ STM m Tick

Get the current " time "

waitUntilDone ∷ m ()

Wait for the end of time (each clock has a maximum number of ticks)

mockSystemTime ∷ SystemTime m

Translate the logical clock to mock SystemTime

newtype NumTicks

Number of ticks the test will run for

Constructors

NumTicks Word64

newtype Tick

Logical time unit

Constructors

Tick

Fields

tickToWord64 ∷ Word64

Construction

new ∷ IOLike m ⇒ ResourceRegistry m → NumTicks → m (LogicalClock m)

sufficientTimeFor ∷ HasCallStack ⇒ [Tick] → NumTicks

Set NumTicks such that we will have seen all of the specified Ticks

Scheduling actions

blockUntilTick ∷ MonadSTM m ⇒ LogicalClock m → Tick → m Bool

Block until the specified tick

Returns False if the current tick is later than the requested one, or True if they were equal.

onTick ∷ (IOLike m, HasCallStack) ⇒ ResourceRegistry m → LogicalClock m → String → Tick → m () → m ()

Execute action once at the specified tick

tickWatcher ∷ LogicalClock m → (Tick → m ()) → Watcher m Tick Tick

Execute action on every clock tick