Test.Util.Orphans.Arbitrary

Documentation

newtype SmallDiffTime

Wrapper around NominalDiffTime with custom Arbitrary instance

The default Arbitrary instance for NominalDiffTime isn't very useful:

It tends to pick huge values

It tends not to pick integer values

It does not shrink

Our custom instance

Picks values between 0 and (1000 * 20 * 10) seconds:

Maximum segment length: 1000

Maximum slot length: 20 seconds

Maximum number of segments: 10

With a 0.1 second precision

Shrinks

Constructors

SmallDiffTime NominalDiffTime

genLimitedEpochSize ∷ Gen EpochSize

The tests for CumulEpochSizes requires that the sum of a list of these values does not overflow.

An epoch size must be > 0.

genLimitedSlotNo ∷ Gen SlotNo

The functions slotAtTime and timeUntilNextSlot suffer from arithmetic overflow for very large values, so generate values that avoid overflow when used in these two functions. The largest value generated is still sufficently large to allow for 5e12 years worth of slots at a slot interval of 20 seconds.

genSmallEpochNo ∷ Gen EpochNo

genSmallSlotNo ∷ Gen SlotNo

Generate a small SlotNo for the state machine tests. The runtime of the StateMachine prop_sequential tests is proportional the the upper bound.

Orphan instances

Arbitrary NominalDiffTime
Picks time span between 0 seconds and (roughly) 50 years

Instance details

Methods

arbitrary ∷ Gen NominalDiffTime

shrink ∷ NominalDiffTime → [NominalDiffTime]

Arbitrary UTCTime
Picks moment between dawnOfTime and (roughly) 50 years later

Uses instance for NominalDiffTime

Instance details

Methods

arbitrary ∷ Gen UTCTime

shrink ∷ UTCTime → [UTCTime]

Arbitrary ClockSkew

Instance details

Methods

arbitrary ∷ Gen ClockSkew

shrink ∷ ClockSkew → [ClockSkew]

Arbitrary Bound

Instance details

Methods

arbitrary ∷ Gen Bound

shrink ∷ Bound → [Bound]

Arbitrary ChunkSlot

Instance details

Methods

arbitrary ∷ Gen ChunkSlot

shrink ∷ ChunkSlot → [ChunkSlot]

Arbitrary ChunkSize
Picks a ChunkSize between 1 and 100, and randomly choose to enable EBBs

Instance details

Methods

arbitrary ∷ Gen ChunkSize

shrink ∷ ChunkSize → [ChunkSize]

Arbitrary ChunkNo
This picks an EpochNo between 0 and 10000

We don't pick larger values because we're not interested in testing overflow due to huge epoch numbers and even huger slot numbers.

Instance details

Methods

arbitrary ∷ Gen ChunkNo

shrink ∷ ChunkNo → [ChunkNo]

Arbitrary RelativeSlot

Instance details

Methods

arbitrary ∷ Gen RelativeSlot

shrink ∷ RelativeSlot → [RelativeSlot]

Arbitrary NumCoreNodes

Instance details

Methods

arbitrary ∷ Gen NumCoreNodes

shrink ∷ NumCoreNodes → [NumCoreNodes]

Arbitrary QueryVersion

Instance details

Methods

arbitrary ∷ Gen QueryVersion

shrink ∷ QueryVersion → [QueryVersion]

Arbitrary SystemStart

Instance details

Methods

arbitrary ∷ Gen SystemStart

shrink ∷ SystemStart → [SystemStart]

Arbitrary SlotLength
Length between 0.001 and 20 seconds, millisecond granularity

Instance details

Methods

arbitrary ∷ Gen SlotLength

shrink ∷ SlotLength → [SlotLength]

Arbitrary BlockNo

Instance details

Methods

arbitrary ∷ Gen BlockNo

shrink ∷ BlockNo → [BlockNo]

(IsNonEmpty xs, SListI xs, All (Compose Arbitrary WrapChainDepState) xs) ⇒ Arbitrary (HardForkChainDepState xs)

Instance details

Methods

arbitrary ∷ Gen (HardForkChainDepState xs)

shrink ∷ HardForkChainDepState xs → [HardForkChainDepState xs]

(All SingleEraBlock (x ': xs), IsNonEmpty xs) ⇒ Arbitrary (MismatchEraInfo (x ': xs))

Instance details

Methods

arbitrary ∷ Gen (MismatchEraInfo (x ': xs))

shrink ∷ MismatchEraInfo (x ': xs) → [MismatchEraInfo (x ': xs)]

Arbitrary (ApplyTxErr blk) ⇒ Arbitrary (WrapApplyTxErr blk)
Forwarding

Instance details

Methods

arbitrary ∷ Gen (WrapApplyTxErr blk)

shrink ∷ WrapApplyTxErr blk → [WrapApplyTxErr blk]

Arbitrary (HeaderHash blk) ⇒ Arbitrary (WrapHeaderHash blk)
Forwarding

Instance details

Methods

arbitrary ∷ Gen (WrapHeaderHash blk)

shrink ∷ WrapHeaderHash blk → [WrapHeaderHash blk]

Arbitrary (TipInfo blk) ⇒ Arbitrary (WrapTipInfo blk)
Forwarding

Instance details

Methods

arbitrary ∷ Gen (WrapTipInfo blk)

shrink ∷ WrapTipInfo blk → [WrapTipInfo blk]

Arbitrary (ChainDepState (BlockProtocol blk)) ⇒ Arbitrary (WrapChainDepState blk)
Forwarding

Instance details

Methods

arbitrary ∷ Gen (WrapChainDepState blk)

shrink ∷ WrapChainDepState blk → [WrapChainDepState blk]

(IsNonEmpty xs, SListI xs, All (Compose Arbitrary LedgerState) xs) ⇒ Arbitrary (LedgerState (HardForkBlock xs))

Instance details

Methods

arbitrary ∷ Gen (LedgerState (HardForkBlock xs))

shrink ∷ LedgerState (HardForkBlock xs) → [LedgerState (HardForkBlock xs)]

SingleEraBlock blk ⇒ Arbitrary (SingleEraInfo blk)

Instance details

Methods

arbitrary ∷ Gen (SingleEraInfo blk)

shrink ∷ SingleEraInfo blk → [SingleEraInfo blk]

SingleEraBlock blk ⇒ Arbitrary (LedgerEraInfo blk)

Instance details

Methods

arbitrary ∷ Gen (LedgerEraInfo blk)

shrink ∷ LedgerEraInfo blk → [LedgerEraInfo blk]

Arbitrary a ⇒ Arbitrary (I a)
Forwarding

Instance details

Methods

arbitrary ∷ Gen (I a)

shrink ∷ I a → [I a]

Arbitrary a ⇒ Arbitrary (WithOrigin a)

Instance details

Methods

arbitrary ∷ Gen (WithOrigin a)

shrink ∷ WithOrigin a → [WithOrigin a]

Arbitrary (f blk) ⇒ Arbitrary (Current f blk)

Instance details

Methods

arbitrary ∷ Gen (Current f blk)

shrink ∷ Current f blk → [Current f blk]

Arbitrary (SomeSecond BlockQuery blk) ⇒ Arbitrary (SomeSecond Query blk)

Instance details

Methods

arbitrary ∷ Gen (SomeSecond Query blk)

shrink ∷ SomeSecond Query blk → [SomeSecond Query blk]

(All (Compose Arbitrary f) xs, IsNonEmpty xs) ⇒ Arbitrary (NS f xs)

Instance details

Methods

arbitrary ∷ Gen (NS f xs)

shrink ∷ NS f xs → [NS f xs]

Arbitrary (K Past blk)

Instance details

Methods

arbitrary ∷ Gen (K Past blk)

shrink ∷ K Past blk → [K Past blk]

(IsNonEmpty xs, All (Compose Arbitrary f) (x ': xs), All (Compose Arbitrary g) (x ': xs)) ⇒ Arbitrary (Mismatch f g (x ': xs))

Instance details

Methods

arbitrary ∷ Gen (Mismatch f g (x ': xs))

shrink ∷ Mismatch f g (x ': xs) → [Mismatch f g (x ': xs)]

(IsNonEmpty xs, All (Compose Arbitrary f) xs, All (Compose Arbitrary g) xs) ⇒ Arbitrary (Telescope g f xs)

Instance details

Methods

arbitrary ∷ Gen (Telescope g f xs)

shrink ∷ Telescope g f xs → [Telescope g f xs]