Test.Util.QuickCheck

Description

QuickCheck utilities

Generic QuickCheck utilities

checkGenerator ∷ (Arbitrary a, Show a) ⇒ (a → Property) → Property

Test the generator

Uses explicit forAll as we don't want to assume a correct shrinker.

checkInvariant ∷ (a → Except String ()) → a → Property

Check invariant

checkShrinker ∷ ∀ a. (Arbitrary a, Show a) ⇒ (a → Property) → Property

Test the shrinker

Comparison functions

expectRight ∷ (Show a, Show b, Eq b) ⇒ b → Either a b → Property

Check that we have the expected Right value

expectRight b ab is roughly equivalent to Right b === ab, but avoids an equality constraint on a.

ge ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like >=, but prints a counterexample when it fails.

gt ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like >, but prints a counterexample when it fails.

le ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like <=, but prints a counterexample when it fails.

lt ∷ (Ord a, Show a) ⇒ a → a → Property infix 4

Like <, but prints a counterexample when it fails.

strictlyIncreasing ∷ ∀ a. (Show a, Ord a) ⇒ [a] → Property

Comparing maps

isSubmapOfBy ∷ (Ord k, Show k, Show a, Show b) ⇒ (a → b → Property) → Map k a → Map k b → Property

Improved variants

elements ∷ HasCallStack ⇒ [a] → Gen a

Generates one of the given values. The input list must be non-empty.

NOTE unlike the standard elements, this variant has a HasCallStack constraint, which makes debugging the error much easier.

(=:=) ∷ (Eq a, Condense a) ⇒ a → a → Property infix 4

Like ===, but uses Condense instead of Show when it fails.

SOP

cshrinkNP ∷ ∀ proxy c f g xs. All c xs ⇒ proxy c → (∀ a. c a ⇒ f a → g a) → (∀ a. c a ⇒ f a → [g a]) → NP f xs → [NP g xs]

shrinkNP ∷ (∀ a. f a → g a) → (∀ a. f a → [g a]) → NP f xs → [NP g xs]

Convenience

collects ∷ Show a ⇒ [a] → Property → Property