Test.Util.RefEnv

Documentation

data RefEnv k a r

elems ∷ RefEnv k a r → [a]

empty ∷ RefEnv k a r

Empty environment

In most context mempty can be used, but the Monoid instance requires equality, which empty does not.

filter ∷ (a → Bool) → RefEnv k a r → RefEnv k a r

fromList ∷ (Eq k, Show k, Eq a, Show a, Eq1 r, Show1 r, HasCallStack) ⇒ [(Reference k r, a)] → RefEnv k a r

keys ∷ RefEnv k a r → [Reference k r]

lookup ∷ (Eq k, Eq1 r) ⇒ Reference k r → RefEnv k a r → Maybe a

null ∷ RefEnv k a r → Bool

reverseLookup ∷ (a → Bool) → RefEnv k a r → [Reference k r]

singleton ∷ Reference k r → a → RefEnv k a r

toList ∷ RefEnv k a r → [(Reference k r, a)]

union ∷ (Eq k, Show k, Eq a, Show a, Eq1 r, Show1 r, HasCallStack) ⇒ RefEnv k a r → RefEnv k a r → RefEnv k a r

(!) ∷ (Eq k, Eq1 r) ⇒ RefEnv k a r → Reference k r → a