Test.Util.Schedule

Description

Utilities to schedule actions per Tick.

Documentation

newtype Schedule a

A schedule plans actions on certain times.

TODO Note that a schedule can't express delays between the actions within a single tick. Generating such delays may expose more (most likely concurrency-related) bugs.

Constructors

Schedule

Fields

getSchedule ∷ Map Tick [a]

genSchedule ∷ [a] → Gen (Schedule a)

Spread out elements over a schedule, i.e. schedule a number of elements to be processed on each tick. Most ticks will have no associated elements.

joinSchedule ∷ Schedule a → [a]

Inverse of genSchedule

lastTick ∷ Schedule a → Tick

Return the last tick at which an update is planned, if no updates are planned, return 0.

shrinkSchedule ∷ Schedule a → [Schedule a]

Repeatedly remove the last entry (highest Tick)