Test.Util.Split

Documentation

spanLeft ∷ ∀ x a b. (x → Either a b) → [x] → ([a], Maybe (b, [x]))

The returned b is the first in the list.

INVARIANT The output data is a segmentation of the given list.

spanLeft' ∷ ∀ x a b. (x → Either a b) → [x] → ([a], Maybe (b, [x]))

As spanLeft, but the [a] is reversed.

splitAtJust ∷ ∀ x b. (x → Maybe b) → Word64 → [x] → (Maybe ([x], b), [x])

The returned b is either the nth b or else the last in the given list.

INVARIANT The output data is a segmentation of the given list.