Test.Util.TestBlock

Description

Minimal instantiation of the consensus layer to be able to run the ChainDB

Blocks

data TestHash where

Bundled Patterns

pattern TestHash ∷ NonEmpty Word64 → TestHash

unTestHash ∷ TestHash → NonEmpty Word64

testHashFromList ∷ [Word64] → TestHash

data TestBlock

Constructors

TestBlock

Fields

tbHash ∷ TestHash

tbSlot ∷ SlotNo

We store a separate SlotNo, as slots can have gaps between them, unlike block numbers.

Note that when generating a TestBlock, you must make sure that blocks with the same TestHash have the same slot number.

tbValid ∷ Bool

Note that when generating a TestBlock, you must make sure that blocks with the same TestHash have the same value for tbValid.

data TestBlockError

Constructors

InvalidHash
The hashes don't line up

Fields

(ChainHash TestBlock)

Expected hash

(ChainHash TestBlock)

Invalid hash

InvalidBlock
The block itself is invalid

data family Header blk

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

firstBlock ∷ Word64 → TestBlock

successorBlock ∷ TestBlock → TestBlock

modifyFork ∷ (Word64 → Word64) → TestBlock → TestBlock

forkBlock ∷ TestBlock → TestBlock

Chain

newtype BlockChain

Constructors

BlockChain Word64

blockChain ∷ BlockChain → Chain TestBlock

chainToBlocks ∷ BlockChain → [TestBlock]

Tree

newtype BlockTree

Constructors

BlockTree (Tree ())

blockTree ∷ BlockTree → Tree TestBlock

treeToBlocks ∷ BlockTree → [TestBlock]

treeToChains ∷ BlockTree → [Chain TestBlock]

treePreferredChain ∷ BlockTree → Chain TestBlock

Ledger infrastructure

lastAppliedBlock ∷ LedgerState TestBlock → Maybe TestBlock

Last applied block

Returns Nothing if the ledger is empty.

testInitLedger ∷ LedgerState TestBlock

testInitExtLedger ∷ ExtLedgerState TestBlock

singleNodeTestConfig ∷ TopLevelConfig TestBlock

Trivial test configuration with a single core node

singleNodeTestConfigWithK ∷ SecurityParam → TopLevelConfig TestBlock

Support for tests

newtype Permutation

Constructors

Permutation Int

permute ∷ Permutation → [a] → [a]