Test.Util.BoolProps

Documentation

class CollectReqs a where

Collect all of the Prereqs and Requirements from a composite type

The default definition uses GHC.Generics to automatically handle algebraic types containing only Prereq, Requirement, Bool, lists, and small tuples.

Note that collectReqs ignores Bools. It's up to the user to interpret Bools as either disjunctive/conjunctive/etc observations.

Minimal complete definition

Nothing

Methods

collectReqs ∷ a → ([Prereq], [Requirement])

default collectReqs ∷ (Generic a, GCollectReqs (Rep a)) ⇒ a → ([Prereq], [Requirement])

data Prereq

A prerequisite

If any prereq is Blocked, then the ultimate observation must be False.

See Requirement.

Constructors

Blocked

Enabled

data Requirement

A requirement

If all prereqs are Enabled and all reqs are Required then the ultimate observation must be True.

Constructors

Optional

Required

checkReqs ∷ CollectReqs a ⇒ a → Maybe Bool

Via CollectReqs, check if the ultimate observation has a required value

enabledIf ∷ Bool → Prereq

gCollectReqs ∷ GCollectReqs rep ⇒ rep (x ∷ Type) → ([Prereq], [Requirement])

requiredIf ∷ Bool → Requirement