Test.Util.TestBlock

Description

Minimal instantiation of the consensus layer to be able to run the ChainDB

Blocks

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family Header blk

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data TestBlockError ptype

Constructors

InvalidHash
The hashes don't line up

Fields

(ChainHash (TestBlockWith ptype))

Expected hash

(ChainHash (TestBlockWith ptype))

Invalid hash

InvalidBlock
The block itself is invalid

InvalidPayload (PayloadDependentError ptype)

data TestBlockWith ptype

Test block parametrized on the payload type

For blocks without payload see the TestBlock type alias.

By defining a PayloadSemantics it is possible to obtain an ApplyBlock instance. See the former class for more details.

data TestHash where

Bundled Patterns

pattern TestHash ∷ NonEmpty Word64 → TestHash

firstBlockWithPayload ∷ Word64 → ptype → TestBlockWith ptype

Create the first block in the given fork, [fork], with the given payload. The SlotNo will be 1.

forkBlock ∷ TestBlock → TestBlock

modifyFork ∷ (Word64 → Word64) → TestBlock → TestBlock

successorBlockWithPayload ∷ TestHash → SlotNo → ptype → TestBlockWith ptype

Create the successor of the given block without forking: b -> b ++ [0] (in the printed representation) The SlotNo is increased by 1.

In Zipper parlance, this corresponds to going down in a tree.

testHashFromList ∷ [Word64] → TestHash

unTestHash ∷ TestHash → NonEmpty Word64

Test block without payload

type TestBlock = TestBlockWith ()

Block without payload

firstBlock ∷ Word64 → TestBlock

See firstBlockWithPayload.

successorBlock ∷ TestBlock → TestBlock

See successorBlockWithPayload.

Payload semantics

class (Typeable ptype, Eq ptype, NoThunks ptype, Eq (PayloadDependentState ptype), Show (PayloadDependentState ptype), Generic (PayloadDependentState ptype), ToExpr (PayloadDependentState ptype), Serialise (PayloadDependentState ptype), NoThunks (PayloadDependentState ptype), Eq (PayloadDependentError ptype), Show (PayloadDependentError ptype), Generic (PayloadDependentError ptype), ToExpr (PayloadDependentError ptype), Serialise (PayloadDependentError ptype), NoThunks (PayloadDependentError ptype), NoThunks (CodecConfig (TestBlockWith ptype)), NoThunks (StorageConfig (TestBlockWith ptype))) ⇒ PayloadSemantics ptype where

Associated Types

type PayloadDependentState ptype ∷ Type

type PayloadDependentError ptype ∷ Type

Methods

applyPayload ∷ PayloadDependentState ptype → ptype → Either (PayloadDependentError ptype) (PayloadDependentState ptype)

LedgerState

lastAppliedPoint ∷ LedgerState (TestBlockWith ptype) → Point (TestBlockWith ptype)

Chain

newtype BlockChain

Constructors

BlockChain Word64

blockChain ∷ BlockChain → Chain TestBlock

chainToBlocks ∷ BlockChain → [TestBlock]

Tree

newtype BlockTree

Constructors

BlockTree (Tree ())

blockTree ∷ BlockTree → Tree TestBlock

treePreferredChain ∷ BlockTree → Chain TestBlock

treeToBlocks ∷ BlockTree → [TestBlock]

treeToChains ∷ BlockTree → [Chain TestBlock]

Ledger infrastructure

singleNodeTestConfig ∷ TopLevelConfig TestBlock

Trivial test configuration with a single core node

singleNodeTestConfigWith ∷ CodecConfig (TestBlockWith ptype) → StorageConfig (TestBlockWith ptype) → SecurityParam → TopLevelConfig (TestBlockWith ptype)

singleNodeTestConfigWithK ∷ SecurityParam → TopLevelConfig TestBlock

testInitExtLedger ∷ ExtLedgerState TestBlock

testInitExtLedgerWithState ∷ PayloadDependentState ptype → ExtLedgerState (TestBlockWith ptype)

testInitLedger ∷ LedgerState TestBlock

testInitLedgerWithState ∷ PayloadDependentState ptype → LedgerState (TestBlockWith ptype)

Support for tests

newtype Permutation

Constructors

Permutation Int

permute ∷ Permutation → [a] → [a]